requirements.txt
kopf
#pykube-ng
pyyaml
azure-mgmt-dns
azure.identity
